requirements.txt
python-periphery==2.0.1
filelock==3.0.12
note-python==1.2.1
websockets==8.1
numpy>=1.19
opencv-python-headless==4.5.1.48
six==1.15.0
psutil==5.8.0
edge_impulse_linux